Implement parenthesized expressions
This required some fiddling with the main expression parse loop, to
accommodate the behavior of the new lparen_prefix vtable function,
which is the heart of parenthesized expression parsing.

`lparen_prefix` calls `expression` straight away, which at first
resulted in an infinite loop because `expression` called `as_prefix`
before advancing through the lexer. This was a bug—I’m not sure why I
wrote it that way originally, but it doesn’t match Pratt’s original
algorithm, which does advance the lexer before calling ‘nud’
(`as_prefix`). So that had to change.

Pretty awesome that we can implement paren exprs without adding a
vtable entry for r-parens, and without having to give l-parens a high
binding power (or any binding power at all, for that matter). Treating
l-parens as a prefix to an expression is pretty cool.