Add lexer_consumef, which exit program
Often want to check `lexer_consume` and exit program if token isn’t
what expected (syntax error).

This isn’t ideal at all—would prefer to return NULL and handle errors
at the very top of the program with a single method—but this ought to
make that refactor a little easier when I get around to it.