Refac ID string handling in lexer, parser
- Clarifies the fact that non-keyword strings are always allocated on
the heap by the lexer, with the assumption that the parser will insert
it into an AST node somewhere. Thus we should never need `strdup` in
the parser.
- Removes the ‘set_tok_val’ functions which weren’t doing enough to be
moved into their own functions and also complicated `consume_id`
somewhat, which doesn’t need to put ID strings on the heap if the ID is
a keyword.