Implement blocks (sequences of statements/expressions)
A LilC program is now parsed as a single top-level block that contains
a series of statements/expressions.

Eventually blocks will be used as the bodies of functions (which
currently are single expressions) and conditional statements.

In codegen, blocks currently evaluate to the value of the last
statement/expression within them. This might be a horrible
decision—we’ll see—but it’s the most straightforward possible
implementation at the moment. It might not end up mattering at all.

This also take on kvec as a dependency as a dynamic array for block
statements. This involved my modifications to CMakeLists.txt.